# pref: solved TODO: improve performance by using matmul33