# TODO: improve performance by using matmul33
At `crates/kornia-icp/src/ops.rs` inside function `fit_transformation` at [line 39](https://github.com/kornia/kornia-rs/blob/6ec2cca5285a1dcc5acd62c55882591c4eca8d6c/crates/kornia-icp/src/ops.rs#L39)

Currently, the ICP implementation uses `faer::linalg::matmul::matmul` for multiplying 3×3 matrices. While `faer` is optimized for general-purpose matrix operations, it's overkill for small fixed-size matrices like 3×3 and introduces unnecessary overhead.

So TODO was to replace this general purpose `matmul` with custom implementation 
Therefore, I added a function `matmul_faer_mat` in `kornia_3d::linalg` that takes 3x3 matrices of type `faer::Mat` and return an mutable pre-allocated matrix that is multiplication of both matrices.

## Implementation pattern:
```rust
 m.write(0, 0, a.read(0, 0) * b.read(0, 0) + a.read(0, 1) * b.read(1, 0) + a.read(0, 2) * b.read(2, 0));
```

## Results: 
Duration it took using my implementation i.e. `matmul_faer_mat`:
```console
233ns
123ns
496ns
35ns
351ns
48ns
34ns
```

Duration it took using `faer::linalg::matmul::matmul`:
```console
353ns
371ns
570ns
193ns
179ns
171ns
180ns
```

## My implementation took **~100ns** less than `faer::linalg::matmul::matmul`