@dhruvkjain 2025-04-07
> ideally we use `glam-rs` for this. How did you get this timings ?

I used `let now = std::time::Instant::now();`, because I saw a similar approach used at `kornia-rs/crates/kornia-icp/src/icp_vanilla.rs` [line 79](https://github.com/kornia/kornia-rs/blob/6ec2cca5285a1dcc5acd62c55882591c4eca8d6c/crates/kornia-icp/src/icp_vanilla.rs#L79), with different types of matrices (Rank-deficient, symmetric, skew symmetric etc.) as inputs.

About `glam-rs`, 
I read it's docs and found `DMat3` matrices multiplication: multiply two 3x3 DMat3 type matrices of f64 values using [`*` operator](https://docs.rs/glam/latest/glam/f64/struct.DMat3.html#method.mul-4) or [`mat_mul3` function](https://docs.rs/glam/latest/glam/f64/struct.DMat3.html#method.mul_mat3).

I feel like adding glam would increase a dependency and would require us to handle new `DMat3` type matrix, my implementation seems pretty straightforward.
I'm not very familiar with `glam-rs`, so I'd appreciate it if you could explain why `glam-rs` might be a better option.