@edgarriba 2025-05-04
glam is a very light dependency with nice and clean abstractions for small matrices operators including optimizations which we'll add in any case as dependency in the 3d crate