@dhruvkjain 2025-05-13
Sorry for late reply,
after doing a bit research how faer, glam and raw implementation would work,
I think for now raw implementation might be better because
- [https://docs.rs/glam/latest/glam/f64/struct.DVec3.html#associatedconstant.USES_WASM32_SIMD](https://docs.rs/glam/latest/glam/f64/struct.DVec3.html#associatedconstant.USES_WASM32_SIMD) : that support SIMD only on wasm32 but rerun uses native platforms for that so even using glam particular for this case wouldn't improve performance 
- and as you already said faer is not the right thing to use for this use case